Add tags work from my time on an airplane.
This introduces basic functionality for DjangoTags and Tags.
For now DjangoTags aren't utilized, but the plan there is to make
them assigned rather than derived. Tags instead will be derived
from the content. In all likelihood, the only tag that will be
used when starting is Django. The remainder of the tags will
need to be reviewed for their relevancy.